TestButtons moved to separate component and inside header bar
